+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 ---
 * Added `Control.Lens.Zipper`
 * Added `<<~`
+* Added `:->` and `:=>` as type operator aliases for `Simple Lens` and `Simple Traversal` respectively
 
 2.9
 ---